Fog/requirements.txt
numpy
matplotlib
scipy
django
setuptools
djangorestframework
drf-yasg